[NSXServiceAccount] Fix incorrect GC in restore case (#351)
Add UID check when creating/deleting NSXServiceAccount to avoid unexpected
PI/CCP/Secret operation.
After NSXT restore finished, PI/CCP will be restored to old state, and it
may contains same cluster name if a cluster is deleted and re-created after
backup point. UID validation is required on this case.

+++ pkg/nsx/services/nsxserviceaccount/cluster.go
@@ -208,7 +208,8 @@ func (s *NSXServiceAccountService) UpdateRealizedNSXServiceAccount(ctx context.C
 
 func (s *NSXServiceAccountService) createPIAndCCP(normalizedClusterName string, vpcPath string, cert string, existingClusterId *string, obj *v1alpha1.NSXServiceAccount) (string, error) {
 	// create PI
-	if piObj := s.PrincipalIdentityStore.GetByKey(normalizedClusterName); piObj == nil {
+	hasPI := len(s.PrincipalIdentityStore.GetByIndex(common.TagScopeNSXServiceAccountCRUID, string(obj.UID))) > 0
+	if piObj := s.PrincipalIdentityStore.GetByKey(normalizedClusterName); !hasPI && piObj == nil {
 		pi, err := s.NSXClient.WithCertificateClient.Create(mpmodel.PrincipalIdentityWithCertificate{
 			IsProtected: &isProtectedTrue,
 			Name:        &normalizedClusterName,
@@ -232,11 +233,14 @@ func (s *NSXServiceAccountService) createPIAndCCP(normalizedClusterName string, 
 			return "", err
 		}
 		s.PrincipalIdentityStore.Add(pi)
+	} else if !hasPI != (piObj == nil) {
+		return "", fmt.Errorf("old PI exists")
 	}
 
 	// create ClusterControlPlane
+	hasCCP := len(s.ClusterControlPlaneStore.GetByIndex(common.TagScopeNSXServiceAccountCRUID, string(obj.UID))) > 0
 	clusterId := ""
-	if ccpObj := s.ClusterControlPlaneStore.GetByKey(normalizedClusterName); ccpObj == nil {
+	if ccpObj := s.ClusterControlPlaneStore.GetByKey(normalizedClusterName); !hasCCP && ccpObj == nil {
 		ccp, err := s.NSXClient.ClusterControlPlanesClient.Update(siteId, enforcementpointId, normalizedClusterName, model.ClusterControlPlane{
 			Revision:     &revision1,
 			ResourceType: &antreaClusterResourceType,
@@ -250,6 +254,8 @@ func (s *NSXServiceAccountService) createPIAndCCP(normalizedClusterName string, 
 		}
 		s.ClusterControlPlaneStore.Add(ccp)
 		clusterId = *ccp.NodeId
+	} else if !hasCCP != (ccpObj == nil) {
+		return "", fmt.Errorf("old CCP exists")
 	}
 	return clusterId, nil
 }
@@ -284,27 +290,45 @@ func (s *NSXServiceAccountService) getProxyEndpoints(ctx context.Context) (v1alp
 	return proxyEndpoints, nil
 }
 
-func (s *NSXServiceAccountService) DeleteNSXServiceAccount(ctx context.Context, namespacedName types.NamespacedName) error {
+func (s *NSXServiceAccountService) DeleteNSXServiceAccount(ctx context.Context, namespacedName types.NamespacedName, uid types.UID) error {
+	isDeleteSecret := false
+	nsxsa := &v1alpha1.NSXServiceAccount{}
+	if err := s.Client.Get(ctx, namespacedName, nsxsa); err != nil {
+		isDeleteSecret = true
+	} else if uid == nsxsa.UID {
+		isDeleteSecret = true
+	}
+
 	clusterName := s.getClusterName(namespacedName.Namespace, namespacedName.Name)
 	normalizedClusterName := util.NormalizeId(clusterName)
 	// delete Secret
-	secretName := namespacedName.Name + SecretSuffix
-	secretNamespace := namespacedName.Namespace
-	if err := s.Client.Delete(ctx, &v1.Secret{ObjectMeta: metav1.ObjectMeta{Name: secretName, Namespace: secretNamespace}}); err != nil && !errors.IsNotFound(err) {
-		log.Error(err, "failed to delete", "secret", secretName, "namespace", secretNamespace)
-		return err
+	if isDeleteSecret {
+		secretName := namespacedName.Name + SecretSuffix
+		secretNamespace := namespacedName.Namespace
+		if err := s.Client.Delete(ctx, &v1.Secret{ObjectMeta: metav1.ObjectMeta{Name: secretName, Namespace: secretNamespace}}); err != nil && !errors.IsNotFound(err) {
+			log.Error(err, "failed to delete", "secret", secretName, "namespace", secretNamespace)
+			return err
+		}
 	}
 
-	// delete ClusterControlPlane
-	cascade := true
-	if err := s.NSXClient.ClusterControlPlanesClient.Delete(siteId, enforcementpointId, normalizedClusterName, &cascade); err != nil {
-		log.Error(err, "failed to delete", "ClusterControlPlane", normalizedClusterName)
-		return err
+	isDeleteCCP := true
+	isDeletePI := true
+	if !isDeleteSecret {
+		isDeletePI = len(s.PrincipalIdentityStore.GetByIndex(common.TagScopeNSXServiceAccountCRUID, string(uid))) > 0
+		isDeleteCCP = len(s.ClusterControlPlaneStore.GetByIndex(common.TagScopeNSXServiceAccountCRUID, string(uid))) > 0
 	}
-	s.ClusterControlPlaneStore.Delete(model.ClusterControlPlane{Id: &normalizedClusterName})
+	// delete ClusterControlPlane
+	if isDeleteCCP {
+		cascade := true
+		if err := s.NSXClient.ClusterControlPlanesClient.Delete(siteId, enforcementpointId, normalizedClusterName, &cascade); err != nil {
+			log.Error(err, "failed to delete", "ClusterControlPlane", normalizedClusterName)
+			return err
+		}
+		s.ClusterControlPlaneStore.Delete(model.ClusterControlPlane{Id: &normalizedClusterName})
+	}
 
 	// delete PI
-	if piobj := s.PrincipalIdentityStore.GetByKey(normalizedClusterName); piobj != nil {
+	if piobj := s.PrincipalIdentityStore.GetByKey(normalizedClusterName); isDeletePI && (piobj != nil) {
 		pi := piobj.(mpmodel.PrincipalIdentity)
 		if err := s.NSXClient.PrincipalIdentitiesClient.Delete(*pi.Id); err != nil {
 			log.Error(err, "failed to delete", "PrincipalIdentity", *pi.Name)
